ISSUE #2087 - Move RiskIcon to separate component
